src/utils/transform-objects-array-to-object.ts
/*
 * Copyright 2020 EPAM Systems
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export const transformObjectsArrayToObject = <T, K extends keyof T>(
  array: T[],
  key: T[K] extends string ? K : never,
) => array.reduce((acc, value) => ({ ...acc, [value[key] as any]: value }), {} as { [key: string]: T });
